+++ examples/Main.hs
@@ -10,9 +10,27 @@ import ExampleEvents
 import ExampleFeatures1
 import ExampleFeatures2
 import ExampleFeatures3
+import ExampleCohort1
+import Test.Tasty
+import Test.Tasty.Hspec
 import Test.Hspec ( hspec )
 
-main :: IO ()
-main = hspec $ do exampleFeatures1Spec
-                  exampleFeatures2Spec
-                  exampleFeatures3Spec 
+-- main = defaultMain tests
+
+tests :: TestTree
+tests = testGroup "Tests" [exampleCohort1tests]
+
+main = do
+  spec <- testSpec "spec" exampleFeatures1Spec
+  defaultMain
+    (testGroup "tests"
+      [ spec
+      , tests
+      ])
+
+
+-- main :: IO ()
+-- main = hspec $ do exampleFeatures1Spec
+--                   exampleFeatures2Spec
+--                   exampleFeatures3Spec
+--                   exampleCohort1tests