+++ test/Hasklepias/AesonSpec.hs
@@ -16,7 +16,7 @@ import qualified Data.ByteString.Lazy as B
 
 
 testInputsDay1 :: B.ByteString
-testInputsDay1 = 
+testInputsDay1 =
       "[\"abc\", \"2020-01-01\", \"2020-01-02\", \"Diagnosis\",\
       \[\"someThing\"],\
       \{\"domain\":\"Diagnosis\",\
@@ -28,7 +28,7 @@ testInputsDay1 =
 
 
 testInputsDay2 :: B.ByteString
-testInputsDay2 = 
+testInputsDay2 =
       "[\"def\", \"2020-01-01\", null, \"Diagnosis\",\
       \[\"someThing\"],\
       \{\"domain\":\"Diagnosis\",\
@@ -42,7 +42,7 @@ testInput = testInputsDay1 <> "\n" <> testInputsDay2
 
 testOutDay1 = event (beginerval 1 (fromGregorian 2020 1 1))
                (HC.context ( Just $ UnimplementedDomain () ) (packConcepts ["someThing"]))
-testOutDay2 = event (beginerval 1 (fromGregorian 2020 1 5)) 
+testOutDay2 = event (beginerval 1 (fromGregorian 2020 1 5))
                (HC.context ( Just $ UnimplementedDomain () ) (packConcepts [ "someThing"]))
 
 
@@ -51,8 +51,8 @@ testOutPop = MkPopulation [
         , MkSubject ("def", [testOutDay1, testOutDay2])
       ]
 
-spec :: Spec 
-spec = do 
-    it "a population is parsed" $ 
-       ( parsePopulationDayLines testInput ) `shouldBe` (testOutPop)
+spec :: Spec
+spec = do
+    it "a population is parsed" $
+       parsePopulationDayLines testInput `shouldBe` testOutPop